Fixes and more error checking. Switch to new cublas api (v2) and add more error checking in lanczos.cu Fix d_aaDMatrixV memory allocation for gpus with more memory in lanczos.cu (Comment out not useful for our project functions in gpb.cu)

+++ bindings/gpb.h
@@ -7,4 +7,4 @@
  *
  */
 
-void gpb(const unsigned int* in_image,unsigned int width,unsigned int height, float* border, int* textons, float* orientations, int device_num=0);
+void gpb(const float* in_image,unsigned int width,unsigned int height, float* border, int* textons, float* orientations, int device_num=0);
